refactor(media-detail): move the delete-file action into the file info panel
The "delete file" action lived in the header's more-menu next to "remove from
library". For users who hold the delete permission, move it to the bottom of the
file information panel it acts on, so the destructive file deletion sits with the
file it targets. Remove-from-library stays in the menu.
